Move hooks into separate files
