+++ packages/core/src/lib/read-line.type.mts
@@ -1,0 +1,8 @@
+import * as readline from 'node:readline';
+import MuteStream from 'mute-stream';
+
+export type InquirerReadline = readline.ReadLine & {
+  output: MuteStream;
+  input: NodeJS.ReadableStream;
+  clearLine: (dir: 0 | 1 | -1) => void; // https://nodejs.org/api/readline.html#rlclearlinedir
+};
