+++ packages/core/src/lib/use-ref.mts
@@ -1,0 +1,5 @@
+import { useState } from './use-state.mjs';
+
+export function useRef<Value>(val: Value): { current: Value } {
+  return useState({ current: val })[0];
+}
